removed JSON metadata added h5_helpers for hdf5 methods for dataset init and metadata writing
