+++ src/compass/utils/runconfig.py
@@ -291,7 +291,7 @@ def create_output_paths(sns, bursts):
         os.makedirs(out_dir, exist_ok=True)
 
         fname_stem = f"{burst_id}_{date_str}_{burst.polarization}"
-        h5_path = f"{out_dir}/{fname_stem}.h5"
+        h5_path = f"{out_dir}/{fname_stem}.nc"
 
         scratch_path = f'{product_paths.scratch_path}/{burst_id}/{date_str}'
         os.makedirs(scratch_path, exist_ok=True)
